fix: Page에 children prop제거, 컴포넌트 분리
