+++ src/app/folder/[[...id]]/Main.module.scss
